+++ packages/adapter-netlify/src/shims.js
@@ -1,5 +1,3 @@
-import { install_crypto } from '@sveltejs/kit/install-crypto';
 import { install_fetch } from '@sveltejs/kit/install-fetch';
 
-install_crypto();
 install_fetch();
