+++ packages/kit/package.json
@@ -86,9 +86,6 @@
 		"./hooks": {
 			"import": "./dist/hooks.js"
 		},
-		"./install-crypto": {
-			"import": "./dist/install-crypto.js"
-		},
 		"./install-fetch": {
 			"import": "./dist/install-fetch.js"
 		}
