+++ packages/kit/rollup.config.js
@@ -57,7 +57,6 @@ export default [
 			cli: 'src/cli.js',
 			node: 'src/node.js',
 			hooks: 'src/hooks.js',
-			'install-crypto': 'src/install-crypto.js',
 			'install-fetch': 'src/install-fetch.js'
 		},
 		output: {
