+++ packages/kit/src/core/dev/plugin.js
@@ -5,7 +5,6 @@ import colors from 'kleur';
 import sirv from 'sirv';
 import { respond } from '../../runtime/server/index.js';
 import { install_fetch } from '../../install-fetch.js';
-import { install_crypto } from '../../install-crypto.js';
 import { create_app } from '../create_app/index.js';
 import create_manifest_data from '../create_manifest_data/index.js';
 import { getRequest, setResponse } from '../../node.js';
@@ -33,7 +32,6 @@ export async function create_plugin(config, cwd) {
 
 		configureServer(vite) {
 			install_fetch();
-			install_crypto();
 
 			/** @type {import('types/app').SSRManifest} */
 			let manifest;
