+++ packages/kit/src/core/preview/index.js
@@ -6,7 +6,6 @@ import sirv from 'sirv';
 import { pathToFileURL } from 'url';
 import { getRequest, setResponse } from '../../node.js';
 import { install_fetch } from '../../install-fetch.js';
-import { install_crypto } from '../../install-crypto.js';
 import { SVELTE_KIT, SVELTE_KIT_ASSETS } from '../constants.js';
 
 /** @param {string} dir */
@@ -33,7 +32,6 @@ export async function preview({
 	cwd = process.cwd()
 }) {
 	install_fetch();
-	install_crypto();
 
 	const app_file = resolve(cwd, `${SVELTE_KIT}/output/server/app.js`);
 	const manifest_file = resolve(cwd, `${SVELTE_KIT}/output/server/manifest.js`);
