--- packages/kit/src/install-crypto.js
@@ -1,12 +1,0 @@
-import { webcrypto } from 'crypto';
-
-// exported for dev/preview and node environments
-export function install_crypto() {
-	Object.defineProperties(globalThis, {
-		crypto: {
-			enumerable: true,
-			configurable: true,
-			value: webcrypto
-		}
-	});
-}
